fixed SSR issue, added code for storing the users cookie consents and restoring them
