+++ src/app/shared/cookies/klaro.effects.ts
@@ -1,0 +1,3 @@
+export class KlaroEffects {
+
+}
