+++ src/modules/app/browser-app.module.ts
@@ -21,6 +21,7 @@ import { AuthService } from '../../app/core/auth/auth.service';
 import { Angulartics2RouterlessModule } from 'angulartics2/routerlessmodule';
 import { SubmissionService } from '../../app/submission/submission.service';
 import { StatisticsModule } from '../../app/statistics/statistics.module';
+import { BrowserKlaroService } from '../../app/shared/cookies/browser-klaro.service';
 import { KlaroService } from '../../app/shared/cookies/klaro.service';
 import { HardRedirectService } from '../../app/core/services/hard-redirect.service';
 import { BrowserHardRedirectService } from '../../app/core/services/browser-hard-redirect.service';
@@ -86,7 +87,7 @@ export function locationProvider(): Location {
     },
     {
       provide: KlaroService,
-      useClass: KlaroService
+      useClass: BrowserKlaroService
     },
     {
       provide: SubmissionService,
