+++ tsconfig.server.json
@@ -6,15 +6,5 @@
   },
   "angularCompilerOptions": {
     "entryModule": "./src/modules/app/server-app.module#ServerAppModule"
-  },
-  "exclude": [
-    "src/test.ts",
-    "src/**/*.spec.ts",
-    "src/**/*.mock.ts",
-    "src/**/*.test.ts",
-    "src/**/*.stub.ts",
-    "src/**/testing/*",
-    "src/**/mocks/*",
-    "node_modules/klaro/*"
-  ],
+  }
 }
